requirements.txt
yt-dlp
opencv-python
moviepy
openai-whisper
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
pydantic
pydantic_settings
seaborn
koreanize_matplotlib
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
plotly
scikit-learn
kiwipiepy
wordcloud
psycopg2-binary
alembic
kaleido